Simplify project identity path calculation
There are 5 implementations of BuildState, each of which implements a method to determine
how a project identity path is derived from a project path.

Ever since #10998, Gradle identity paths are known statically and are identical to the BuildState identity paths.
Therefore the nested builds which used getBuildController().getGradle().getIdentityPath() as the base of their
project paths can just use their own identity path as base.